renamed main file

+++ src/pecuzal_method.jl
@@ -78,7 +78,7 @@ For distance computations the Euclidean norm is used.
 """
 function pecuzal_embedding(s::Vector{T}; τs = 0:50 , w::Int = 1,
     samplesize::Real = 1, K::Int = 13, KNN::Int = 3, Tw::Int=4*w,
-    α::Real = 0.05, p::Real = 0.5,max_cycles = 50) where {T<:Real}
+    α::Real = 0.05, p::Real = 0.5, max_cycles::Int = 50) where {T<:Real}
 
     @assert 0 < samplesize ≤ 1 "Please select a valid `samplesize`, which denotes a fraction of considered fiducial points, i.e. `samplesize` ∈ (0 1]"
     @assert all(x -> x ≥ 0, τs)
@@ -123,7 +123,7 @@ end
 
 function pecuzal_embedding(Y::Dataset{D, T}; τs = 0:50 , w::Int = 1,
     samplesize::Real = 1, K::Int = 13, KNN::Int = 3, Tw::Int=4*w,
-    α::Real = 0.05, p::Real = 0.5, max_cycles = 50) where {D, T<:Real}
+    α::Real = 0.05, p::Real = 0.5, max_cycles::Int = 50) where {D, T<:Real}
 
     @assert 0 < samplesize ≤ 1 "Please select a valid `samplesize`, which denotes a fraction of considered fiducial points, i.e. `samplesize` ∈ (0 1]"
     @assert all(x -> x ≥ 0, τs)
